==> 1. Data Wrangling - Metadata.ipynb <==
# Data Wrangling - Audio Metadata
The audio clips used in this project came from two separate databases with different naming conventions. Before getting into the audio filtering and feature selection, I want to create a csv with all the data for each audio clip neatly combined so that I can use it as a legend of sorts.

Both databases store information in the filename of each clip. In the case for CREMA-D, more information concerning the actors are found in a separate file `CREMA-D Actors.csv`.

By the end of this process, I’ll have a singular file, `Audio Legend.csv` with the following columns:

 - Path: path to the audio clip
 - Original: the original clip filename
 - Statement: an integer denoting which of the fourteen phrases was used
 - Emotion: the emotion portrayed in the clip
 - Sex: the sex of the actor
 - Race: the race of the actor



import numpy as np
import os
import pandas as pd

# Collection
Going into the `Data` folder to select the filenames of each audio clip, breaking their names into the respective components, and saving them into a separate dataframe for each database.

data_RAVDESS = pd.DataFrame()
error = []

folder = os.listdir('Data/Audio/Raw/RAVDESS/')
for count, track in enumerate(folder,start=1):
    path = 'Data/Audio/Raw/RAVDESS/'+track
    try:
        # make a list using the clip name
        metadata = [path[:-len(track)],track]
        # cut up the name into its seperate components and append to the list
        [metadata.append(int(x)) for x in track[:-4].split('-')]
        # add the list to the RAVDESS dataset
        data_RAVDESS = data_RAVDESS.append([metadata])
    except:
        error.append(track)
    # progress bar
    print('%4d of %d complete (%d%%)' % (count,len(folder),(count/len(folder))*100),end='\r')
            
data_RAVDESS.head(3)

data_CREMA_D = pd.DataFrame()

folder = os.listdir('Data/Audio/Raw/CREMA_D/')
for count, track in enumerate(folder,start=1):
    path = 'Data/Audio/Raw/CREMA_D/'+track
    try:
        # make a list using the clip name
        metadata = [path[:-len(track)],track]
        # cut up the name into its seperate components and append to the list
        [metadata.append(x) for x in track[:-4].split('_')]
         # add the list to the CREMA-D dataset
        data_CREMA_D = data_CREMA_D.append([metadata])
    except:
        error.append(track)
    # progress bar
    print('%4d of %d complete (%d%%)' % (count,len(folder),(count/len(folder))*100),end='\r')
        
data_CREMA_D.head(3)

# Cleaning
Clean and combine both datasets

# change index to a range of ints
data_RAVDESS.index = pd.RangeIndex(0,stop = len(data_RAVDESS))
data_CREMA_D.index = pd.RangeIndex(0,stop = len(data_CREMA_D))

# drop unnecessary columns
data_RAVDESS = data_RAVDESS.drop(columns=[2,3,5,7])
data_CREMA_D = data_CREMA_D.drop(columns=[5])

# rename columns
data_RAVDESS.columns=['path','original','emotion','statement','actorid']
data_CREMA_D.columns=['path','original','ActorID','statement','emotion']

# make sure id is an int not an object
data_CREMA_D[['ActorID']] = data_CREMA_D[['ActorID']].astype('int')

data_RAVDESS.info()
data_CREMA_D.info()

### Label Convention
Make sure that both datasets have the same column labels

# Make all the emotions human redable
for i,j in [('ANG','anger'),('DIS','disgust'),('FEA','fear'),('NEU','neutral'),('HAP','happy'),('SAD','sad')]:
    data_CREMA_D['emotion'] = data_CREMA_D['emotion'].str.replace(i,j)
    
for i,j in [(1,'neutral'),(2,'calm'),(3,'happy'),(4,'sad'),(5,'anger'),(6,'fear'),(7,'disgust'),(8,'surprised')]:
    data_RAVDESS['emotion'] = data_RAVDESS['emotion'].replace(i,j)

# Actor gender
CREMA_D = pd.read_csv('Data/CSVs/CREMA_D Actors.csv',index_col='Unnamed: 0')
data_CREMA_D = data_CREMA_D.merge(CREMA_D[['Sex','ActorID','Race','Ethnicity']],on='ActorID',how='inner')
data_CREMA_D.columns = data_CREMA_D.columns.str.lower()

data_RAVDESS['sex'] = data_RAVDESS['actorid'] % 2
data_RAVDESS['sex'] = data_RAVDESS['sex'].replace(1,'Male').replace(0,'Female')

# change the statement convention
for i,j in enumerate(data_CREMA_D['statement'].unique(),start=3):
    data_CREMA_D['statement'] = data_CREMA_D['statement'].replace(j,i)

Overall there are 14 total statements used:

- 1: Kids are talking by the door.
- 2: Dogs are sitting by the door.
- 3: Don't forget a jacket.
- 4: It's eleven o'clock.
- 5: I'm on my way to the meeting.
- 6: I think I have a doctor's appointment.
- 7 I think I've seen this before.
- 8: I would like a new alarm clock.
- 9: I wonder what this is about.
- 10: Maybe tomorrow it will be cold.
- 11: The airplane is almost full.
- 12: That is exactly what happened.
- 13: The surface is slick.
- 14: We'll stop in a couple of minutes.

### Final Check
Make sure the columns and column labels are the same.

data_CREMA_D.head(3)

data_RAVDESS.head(3)

# append data together
data = data_CREMA_D.append(data_RAVDESS)
data.index = pd.RangeIndex(0,len(data))
data.info()

#save as csv
data.to_csv('Data/CSVs/Audio Legend Raw.csv')
data.head()
